# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EV ChargeOps API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/me": {
      "get": {
        "tags": [
          "identity"
        ],
        "summary": "Me",
        "operationId": "me_v1_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Resident-Context",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resident-Context"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/resident-context": {
      "post": {
        "tags": [
          "identity"
        ],
        "summary": "Enter Resident Context",
        "description": "See exactly what this unit's resident would see, for a short while.\n\nThe entrance is audited with actor, unit and instant. The context resolves\nthrough the same authorization dependency a real resident session would\nuse, so every write is rejected while it is active.",
        "operationId": "enter_resident_context_v1_resident_context_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterResidentContextRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "identity"
        ],
        "summary": "Exit Resident Context",
        "description": "One-click exit. The departure is audited like the entrance was.",
        "operationId": "exit_resident_context_v1_resident_context_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Resident-Context",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Resident-Context"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnterResidentContextRequest": {
        "properties": {
          "unitId": {
            "type": "string",
            "format": "uuid",
            "title": "Unitid"
          }
        },
        "type": "object",
        "required": [
          "unitId"
        ],
        "title": "EnterResidentContextRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeResponse": {
        "properties": {
          "profileId": {
            "type": "string",
            "format": "uuid",
            "title": "Profileid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "title": "Organizationid"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole"
          },
          "unitId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unitid"
          }
        },
        "type": "object",
        "required": [
          "profileId",
          "organizationId",
          "role",
          "unitId"
        ],
        "title": "MeResponse"
      },
...
      "OrganizationRole": {
        "type": "string",
        "enum": [
          "manager",
          "resident",
          "technical_operator"
        ],
        "title": "OrganizationRole"
      },
...
      "ResidentContextResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "unitId": {
            "type": "string",
            "format": "uuid",
            "title": "Unitid"
          },
          "unitCode": {
            "type": "string",
            "title": "Unitcode"
          },
          "unitName": {
            "type": "string",
            "title": "Unitname"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "unitId",
          "unitCode",
          "unitName",
          "expiresAt"
        ],
        "title": "ResidentContextResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}